+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Introduction of `GridInspector()` class to investigate properties of a Dataset/DataArray (#33)
 - Minimal documentation via Sphinx available through ReadTheDocs (#33)
 - Preliminar support to target unstructured grids (#33)
+- `CdoGenerate()` class to support weights generator through `weights()` method (#33)
+- `CdoGenerate().areas()` method to generate areas based on CDO (#33)
 
 ## [v0.0.7]